+++ Civi/DataProcessor/Source/AbstractCivicrmEntitySource.php
@@ -6,7 +6,10 @@
 
 namespace Civi\DataProcessor\Source;
 
+use Civi\DataProcessor\DataFlow\CombinedDataFlow\SubqueryDataFlow;
 use Civi\DataProcessor\DataFlow\MultipleDataFlows\JoinInterface;
+use Civi\DataProcessor\DataFlow\MultipleDataFlows\SimpleNonRequiredJoin;
+use Civi\DataProcessor\DataFlow\SqlDataFlow\PureSqlStatementClause;
 use Civi\DataProcessor\DataFlow\SqlDataFlow\SimpleWhereClause;
 use Civi\DataProcessor\DataFlow\SqlTableDataFlow;
 use Civi\DataProcessor\DataFlow\CombinedDataFlow\CombinedSqlDataFlow;
@@ -19,6 +22,8 @@ use Civi\DataProcessor\DataSpecification\FieldSpecification;
 use Civi\DataProcessor\DataSpecification\Utils as DataSpecificationUtils;
 use Civi\DataProcessor\ProcessorType\AbstractProcessorType;
 
+use CRM_Dataprocessor_ExtensionUtil as E;
+
 abstract class AbstractCivicrmEntitySource extends AbstractSource {
 
   /**
@@ -30,6 +35,16 @@ abstract class AbstractCivicrmEntitySource extends AbstractSource {
    * @var \Civi\DataProcessor\DataFlow\SqlDataFlow
    */
   protected $primaryDataFlow;
+
+  /**
+   * @var SqlTableDataFlow
+   */
+  protected $entityDataFlow;
+
+  /**
+   * @var SubqueryDataFlow
+   */
+  protected $aggregationDateFlow;
 
   /**
    * @var \Civi\DataProcessor\DataSpecification\DataSpecification
@@ -135,7 +150,48 @@ abstract class AbstractCivicrmEntitySource extends AbstractSource {
    * @return \Civi\DataProcessor\DataFlow\SqlDataFlow
    */
   protected function getEntityDataFlow() {
-    return new SqlTableDataFlow($this->getTable(), $this->getSourceName());
+    if (!$this->entityDataFlow) {
+      $this->entityDataFlow = new SqlTableDataFlow($this->getTable(), $this->getSourceName());
+    }
+    if ($this->isAggregationEnabled()) {
+      $start_date_field_spec = clone $this->getAvailableFields()->getFieldSpecificationByName($this->getAggregateField());
+      $start_date_field_spec->setMySqlFunction($this->getAggregateFunction());
+      $start_date_field_spec->alias = $this->getAggregateField();
+
+      $groupByFields = array();
+      foreach($this->configuration['aggregate_by'] as $aggregate_by) {
+        $field = clone $this->getAvailableFields()->getFieldSpecificationByName($aggregate_by);
+        $field->alias = $aggregate_by;
+        $groupByFields[] = $field;
+      }
+
+      if (!$this->aggregationDateFlow) {
+        $aggretaed_membership_table_dataflow = new SqlTableDataFlow($this->getTable(), '_aggregated_'.$this->getSourceName());
+        $this->aggregationDateFlow = new SubqueryDataFlow('', $this->getTable(), '_aggregated_'.$this->getSourceName());
+        $this->aggregationDateFlow->addSourceDataFlow(new DataFlowDescription($aggretaed_membership_table_dataflow));
+        $aggretaed_membership_table_dataflow->getDataSpecification()->addFieldSpecification($start_date_field_spec->name, $start_date_field_spec);
+
+        foreach($groupByFields as $groupByField) {
+          $aggretaed_membership_table_dataflow->getDataSpecification()
+            ->addFieldSpecification($groupByField->name, $groupByField);
+          $aggretaed_membership_table_dataflow->getGroupByDataSpecification()
+            ->addFieldSpecification($groupByField->name, $groupByField);
+        }
+      }
+
+      $dataFlow = new CombinedSqlDataFlow('', $aggretaed_membership_table_dataflow->getTable(), $aggretaed_membership_table_dataflow->getTableAlias());
+      $dataFlow->addSourceDataFlow(new DataFlowDescription($this->aggregationDateFlow));
+      $join = new SimpleNonRequiredJoin($this->entityDataFlow->getTableAlias(), $this->getAggregateField(), $aggretaed_membership_table_dataflow->getTableAlias(), $this->getAggregateField(), 'INNER');
+      foreach($groupByFields as $groupByField) {
+        $join->addFilterClause(new PureSqlStatementClause("`{$this->entityDataFlow->getTableAlias()}`.`{$groupByField->alias}` = `{$aggretaed_membership_table_dataflow->getTableAlias()}`.`{$groupByField->alias}`", TRUE));
+      }
+      $join->setDataProcessor($this->dataProcessor);
+      $dataFlowDescription = new DataFlowDescription($this->entityDataFlow, $join);
+      $dataFlow->addSourceDataFlow($dataFlowDescription);
+      return $dataFlow;
+    } else {
+      return $this->entityDataFlow;
+    }
   }
 
   /**
@@ -275,11 +331,11 @@ abstract class AbstractCivicrmEntitySource extends AbstractSource {
    * @throws \Exception
    */
   protected function ensureEntity() {
-    if ($this->primaryDataFlow && $this->primaryDataFlow->getTable() === $this->getTable()) {
-      return $this->primaryDataFlow;
-    } elseif (empty($this->primaryDataFlow)) {
-      $this->primaryDataFlow = $this->getEntityDataFlow();
-      return $this->primaryDataFlow;
+    if ($this->entityDataFlow && $this->entityDataFlow->getTable() === $this->getTable()) {
+      return $this->entityDataFlow;
+    } elseif (empty($this->entityDataFlow)) {
+      $this->getEntityDataFlow();
+      return $this->entityDataFlow;
     }
     foreach($this->additionalDataFlowDescriptions as $additionalDataFlowDescription) {
       if ($additionalDataFlowDescription->getDataFlow()->getTable() == $this->getTable()) {
@@ -287,11 +343,11 @@ abstract class AbstractCivicrmEntitySource extends AbstractSource {
       }
     }
     $entityDataFlow = $this->getEntityDataFlow();
-    $join = new SimpleJoin($this->getSourceName(), 'id', $this->primaryDataFlow->getTableAlias(), 'entity_id', 'LEFT');
+    $join = new SimpleJoin($this->getSourceName(), 'id', $this->entityDataFlow->getTableAlias(), 'entity_id', 'LEFT');
     $join->setDataProcessor($this->dataProcessor);
     $additionalDataFlowDescription = new DataFlowDescription($entityDataFlow,$join);
     $this->additionalDataFlowDescriptions[] = $additionalDataFlowDescription;
-    return $additionalDataFlowDescription->getDataFlow();
+    return $this->entityDataFlow;
   }
 
   /**
@@ -307,6 +363,16 @@ abstract class AbstractCivicrmEntitySource extends AbstractSource {
         $this->primaryDataFlow = $customGroupDataFlowDescription->getDataFlow();
         unset($this->customGroupDataFlowDescriptions[$idx]);
         unset($this->dataFlow);
+      }
+    }
+    if ($this->isAggregationEnabled()) {
+      if ($join instanceof SimpleJoin && $join->getLeftTable() == $this->entityDataFlow->getTableAlias()) {
+        $join->setLeftTable($this->aggregationDateFlow->getPrimaryTableAlias());
+        $join->setLeftPrefix($this->aggregationDateFlow->getPrimaryTableAlias());
+      }
+      elseif ($join instanceof SimpleJoin && $join->getRightTable() == $this->entityDataFlow->getTableAlias()) {
+        $join->setRightTable($this->aggregationDateFlow->getPrimaryTableAlias());
+        $join->setRightPrefix($this->aggregationDateFlow->getPrimaryTableAlias());
       }
     }
     return $this;
@@ -372,6 +438,127 @@ abstract class AbstractCivicrmEntitySource extends AbstractSource {
    */
   public function getPrimaryDataFlow() {
     return $this->primaryDataFlow;
+  }
+
+  /**
+   * Returns true when aggregation is enabled for this data source.
+   *
+   * @return bool
+   */
+  protected function isAggregationEnabled() {
+    if (!empty($this->configuration['aggregate_function'])) {
+      return TRUE;
+    }
+    return false;
+  }
+
+  /**
+   * @return string|false
+   */
+  protected function getAggregateFunction() {
+    if (strpos($this->configuration['aggregate_function'], 'min_') === 0) {
+      return 'MIN';
+    } elseif (strpos($this->configuration['aggregate_function'], 'max_') === 0) {
+      return 'MAX';
+    }
+    return false;
+  }
+
+  /**
+   * Returns the name of the aggregated field.
+   *
+   * @return string|false
+   */
+  protected function getAggregateField() {
+    if (strpos($this->configuration['aggregate_function'], 'min_') === 0) {
+      return substr($this->configuration['aggregate_function'], 4);
+    } elseif (strpos($this->configuration['aggregate_function'], 'max_') === 0) {
+      return substr($this->configuration['aggregate_function'], 4);
+    }
+    return false;
+  }
+
+  /**
+   * Returns an array with possible aggregate functions.
+   * Return false when aggregation is not possible.
+   *
+   * This function could be overridden in child classes.
+   *
+   * @return array|false
+   */
+  protected function getPossibleAggregateFunctions() {
+    return false;
+  }
+
+  /**
+   * When this source has configuration specify the template file name
+   * for the configuration form.
+   *
+   * @return false|string
+   */
+  public function getConfigurationTemplateFileName() {
+    return "CRM/Dataprocessor/Form/Source/CiviCRMEntitySourceConfiguration.tpl";
+  }
+
+  /**
+   * When this source has additional configuration you can add
+   * the fields on the form with this function.
+   *
+   * @param \CRM_Core_Form $form
+   * @param array $source
+   */
+  public function buildConfigurationForm(\CRM_Core_Form $form, $source=array()) {
+    parent::buildConfigurationForm($form, $source);
+    $fields = array();
+    foreach($this->getAvailableFields()->getFields() as $field) {
+      $fields[$field->getName()] = $field->title;
+    }
+    $aggregateFunctions = $this->getPossibleAggregateFunctions();
+    if (is_array($aggregateFunctions)) {
+      $form->add('select', "aggregate_function", E::ts('Aggregate function'), $aggregateFunctions, FALSE, [
+        'style' => 'min-width:250px',
+        'class' => 'crm-select2 huge',
+        'placeholder' => E::ts('No aggregation'),
+      ]);
+      $form->add('select', "aggregate_by", E::ts('Aggregate by'), $fields, FALSE, [
+        'style' => 'min-width:250px',
+        'class' => 'crm-select2 huge',
+        'multiple' => 'multiple',
+        'placeholder' => E::ts('- select -'),
+      ]);
+      $form->addFormRule([$this, 'validateConfigurationForm']);
+      $defaults = array();
+      if (isset($source['configuration']['aggregate_function'])) {
+        $defaults['aggregate_function'] = $source['configuration']['aggregate_function'];
+      }
+      if (isset($source['configuration']['aggregate_by'])) {
+        $defaults['aggregate_by'] = $source['configuration']['aggregate_by'];
+      }
+      $form->setDefaults($defaults);
+    }
+  }
+
+  public function validateConfigurationForm($fields) {
+    $errors = [];
+    if (!empty($fields['aggregate_function']) && empty($fields['aggregate_by'])) {
+      $errors['aggregate_by'] = E::ts('Select at least one field at aggregate by');
+    }
+    return count($errors) ? $errors : true;
+  }
+
+  /**
+   * Process the submitted values and create a configuration array
+   *
+   * @param $submittedValues
+   * @return array
+   */
+  public function processConfiguration($submittedValues) {
+    $configuration = parent::processConfiguration($submittedValues);
+    if (isset($submittedValues['aggregate_function'])) {
+      $configuration['aggregate_function'] = $submittedValues['aggregate_function'];
+      $configuration['aggregate_by'] = $submittedValues['aggregate_by'];
+    }
+    return $configuration;
   }
 
 }